# Add an optimized http parser for using the same ideas as nom-optimized