--- a/http/combine-http/src/main.rs
+++ b/http/combine-http/src/main.rs
@@ -5,11 +5,8 @@ extern crate combine;
 
 use bencher::{black_box, Bencher};
 
-use std::fmt;
-
-use combine::{many, token, ParseError, Parser, RangeStream, many1};
 use combine::range::{range, take_while1};
-use combine::stream::easy;
+use combine::{many, token, ParseError, Parser, RangeStream, many1};
 
 #[derive(Debug)]
 struct Request<'a> {
@@ -63,14 +60,40 @@ fn is_http_version(c: u8) -> bool {
     c >= b'0' && c <= b'9' || c == b'.'
 }
 
+fn end_of_line<'a, I>() -> impl Parser<Output = u8, Input = I>
+where
+    I: RangeStream<Item = u8, Range = &'a [u8]>,
+    I::Error: ParseError<I::Item, I::Range, I::Position>,
+{
+    (token(b'\r'), token(b'\n')).map(|_| b'\r').or(token(b'\n'))
+}
+
+fn message_header<'a, I>() -> impl Parser<Output = Header<'a>, Input = I>
+where
+    I: RangeStream<Item = u8, Range = &'a [u8]>,
+    I::Error: ParseError<I::Item, I::Range, I::Position>,
+{
+    let message_header_line = (
+        take_while1(is_horizontal_space),
+        take_while1(|c| c != b'\r' && c != b'\n'),
+        end_of_line(),
+    ).map(|(_, line, _)| line);
+
+    (
+        take_while1(is_token),
+        token(b':'),
+        many1(message_header_line),
+    ).map(|(name, _, value)| Header {
+        name: name,
+        value: value,
+    })
+}
+
 fn parse_http_request<'a, I>(input: I) -> Result<((Request<'a>, Vec<Header<'a>>), I), I::Error>
 where
     I: RangeStream<Item = u8, Range = &'a [u8]>,
     I::Error: ParseError<I::Item, I::Range, I::Position>,
 {
-    // Making a closure, because parser instances cannot be reused
-    let end_of_line = || (token(b'\r'), token(b'\n')).map(|_| b'\r').or(token(b'\n'));
-
     let http_version = range(&b"HTTP/"[..]).with(take_while1(is_http_version));
 
     let request_line = struct_parser!(Request {
@@ -81,25 +104,10 @@ where
             version: http_version,
         });
 
-    let message_header_line = (
-        take_while1(is_horizontal_space),
-        take_while1(|c| c != b'\r' && c != b'\n'),
-        end_of_line(),
-    ).map(|(_, line, _)| line);
-
-    let message_header = (
-        take_while1(is_token),
-        token(b':'),
-        many1(message_header_line),
-    ).map(|(name, _, value)| Header {
-        name: name,
-        value: value,
-    });
-
     let mut request = (
         request_line,
         end_of_line(),
-        many(message_header),
+        many(message_header()),
         end_of_line(),
     ).map(|(request, _, headers, _)| (request, headers));
 
@@ -107,19 +115,19 @@ where
 }
 
 fn small_test(b: &mut Bencher) {
-  let data = include_bytes!("../../http-requests.txt");
-  b.bytes = data.len() as u64;
-  parse(b, data)
+    let data = include_bytes!("../../http-requests.txt");
+    b.bytes = data.len() as u64;
+    parse(b, data)
 }
 
 fn bigger_test(b: &mut Bencher) {
-  let data = include_bytes!("../../bigger.txt");
-  b.bytes = data.len() as u64;
-  parse(b, data)
+    let data = include_bytes!("../../bigger.txt");
+    b.bytes = data.len() as u64;
+    parse(b, data)
 }
 
 fn one_test(b: &mut Bencher) {
-  let data = &b"GET / HTTP/1.1
+    let data = &b"GET / HTTP/1.1
 Host: www.reddit.com
 User-Agent: Mozilla/5.0 (Macintosh; Intel Mac OS X 10.8; rv:15.0) Gecko/20100101 Firefox/15.0.1
 Accept: text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8
@@ -128,12 +136,12 @@ Accept-Encoding: gzip, deflate
 Connection: keep-alive
 
 "[..];
-  b.bytes = data.len() as u64;
-  parse(b, data)
+    b.bytes = data.len() as u64;
+    parse(b, data)
 }
 
 fn httparse_example_test(b: &mut Bencher) {
-  let data = &b"GET /wp-content/uploads/2010/03/hello-kitty-darth-vader-pink.jpg HTTP/1.1\r\n\
+    let data = &b"GET /wp-content/uploads/2010/03/hello-kitty-darth-vader-pink.jpg HTTP/1.1\r\n\
 Host: www.kittyhell.com\r\n\
 User-Agent: Mozilla/5.0 (Macintosh; U; Intel Mac OS X 10.6; ja-JP-mac; rv:1.9.2.3) Gecko/20100401 Firefox/3.6.3 Pathtraq/0.9\r\n\
 Accept: text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8\r\n\
@@ -144,8 +152,8 @@ Keep-Alive: 115\r\n\
 Connection: keep-alive\r\n\
 Cookie: wp_ozh_wsa_visits=2; wp_ozh_wsa_visit_lasttime=xxxxxxxxxx; __utma=xxxxxxxxx.xxxxxxxxxx.xxxxxxxxxx.xxxxxxxxxx.xxxxxxxxxx.x; __utmz=xxxxxxxxx.xxxxxxxxxx.x.x.utmccn=(referral)|utmcsr=reader.livedoor.com|utmcct=/reader/|utmcmd=referral\r\n\r\n"[..];
 
-  b.bytes = data.len() as u64;
-  parse(b, data)
+    b.bytes = data.len() as u64;
+    parse(b, data)
 }
 
 fn parse(b: &mut Bencher, buffer: &[u8]) {
@@ -169,6 +177,11 @@ fn parse(b: &mut Bencher, buffer: &[u8]) {
     });
 }
 
-benchmark_group!(http, one_test, small_test, bigger_test, httparse_example_test);
+benchmark_group!(
+    http,
+    one_test,
+    small_test,
+    bigger_test,
+    httparse_example_test
+);
 benchmark_main!(http);
-