--- a/http/combine-http/src/main.rs
+++ b/http/combine-http/src/main.rs
@@ -1,5 +1,5 @@
 #![feature(const_fn)]
-#![feature(cfg_target_feature, target_feature, stdsimd)]
+#![feature(stdsimd)]
 
 #[macro_use]
 extern crate bencher;
@@ -8,10 +8,11 @@ extern crate combine;
 
 use bencher::{black_box, Bencher};
 
-use combine::{many, token, ParseError, parser, Parser, RangeStream, many1, skip_many};
-use combine::range::{range, take, take_while1};
-use combine::stream::{FullRangeStream, easy};
+use combine::{token, ParseError, parser, Parser, RangeStream, skip_many};
+use combine::range::{range, take_while1};
+use combine::stream::FullRangeStream;
 use combine::error::Consumed;
+use combine::parser::combinator::no_partial;
 
 
 #[path = "../../nom-optimized/src/combinators.rs"]
@@ -42,9 +43,6 @@ fn is_horizontal_space(c: u8) -> bool {
 fn is_space(c: u8) -> bool {
     c == b' '
 }
-fn is_not_space(c: u8) -> bool {
-    c != b' '
-}
 fn is_http_version(c: u8) -> bool {
     c >= b'0' && c <= b'9' || c == b'.'
 }
@@ -58,7 +56,7 @@ where
 {
     parser(move |input: &mut I| {
         match combinators::take_while1_simd(input.range(), &mut predicate, range) {
-            Ok((remaining, value)) => {
+            Ok((_, value)) => {
                 let _ = input.uncons_range(value.len());
                 Ok((value, Consumed::Consumed(())))
             }
@@ -80,47 +78,47 @@ where
     I: FullRangeStream<Item = u8, Range = &'a [u8]> + 'a,
     I::Error: ParseError<I::Item, I::Range, I::Position>,
 {
-    const header_value_range:&[u8] = b"\0\x08\x0A\x1F\x7F\x7F";
-    let message_header_line = (
+    const HEADER_VALUE_RANGE:&[u8] = b"\0\x08\x0A\x1F\x7F\x7F";
+    let message_header_line = no_partial((
         take_while1(is_horizontal_space),
-        take_while1_simd(header_value_range, is_header_value_token),
+        take_while1_simd(HEADER_VALUE_RANGE, is_header_value_token),
         end_of_line(),
-    ).map(|(_, line, _)| line);
+    )).map(|(_, line, _)| line);
 
-    (
+    no_partial((
         take_while1(is_token),
         token(b':'),
         message_header_line,
-    ).map(|(name, _, value)| Header {
+    )).map(|(name, _, value)| Header {
         name: name,
         value: value,
     })
 }
 
-fn parse_http_request<'a, I>(input: I, mut headers: &mut [Header<'a>]) -> Result<(Request<'a>, I), I::Error>
+fn parse_http_request<'a, I>(input: I, headers: &mut [Header<'a>]) -> Result<(Request<'a>, I), I::Error>
 where
     I: FullRangeStream<Item = u8, Range = &'a [u8]> + 'a,
     I::Error: ParseError<I::Item, I::Range, I::Position>,
 {
     let http_version = range(&b"HTTP/"[..]).with(take_while1(is_http_version));
 
-    let request_line = struct_parser!(Request {
+    let request_line = no_partial(struct_parser!(Request {
             method: take_while1(is_token),
             _: take_while1(is_space),
             uri: take_while1_simd(b"\0 \x7F\x7F", is_url_token),
             _: take_while1(is_space),
             version: http_version,
-        });
+        }));
 
     let mut headers = headers.iter_mut();
-    let mut request = (
+    let mut request = no_partial((
         request_line,
         end_of_line(),
         skip_many(message_header().map(|header| {
             *headers.next().unwrap() = header;
         })),
         end_of_line(),
-    ).map(|(request, _, _, _)| request);
+    )).map(|(request, _, _, _)| request);
 
     request.parse(input)
 }